requirements.txt
streamlit
psycopg2-binary
pandas
pytz
SQLAlchemy
python-dotenv
streamlit-option-menu
plotly
requests
streamlit-autorefresh
streamlit-js-eval
openai>=1.0.0
fastapi
uvicorn
qrcode
Pillow
reportlab
numpy
